# api/openapi.yaml
openapi: 3.1.0
info:
  title: Coop API
  version: 0.1.0
  description: |
    Coop is a self-hosted, parent-curated YouTube for kids.

    This document is the contract between the Go backend and the two native
    clients. The Swift client is generated from it, and the server is tested
    against it so drift fails before release rather than at runtime.

    Two disjoint surfaces share one server:

    - `/api/v1/parent` requires a parent session. Parents scoped to specific
      children receive 404 rather than 403 for children outside their scope, so
      the response does not confirm that another child exists.
    - `/api/v1/child` requires a paired device token. It is read-mostly: a child
      can read their own feed, record local reactions and watch progress, and
      raise a request. Nothing on this surface can change policy.
  license:
    name: MIT
    url: https://github.com/nerdswhofish/coop/blob/main/LICENSE
servers:
  - url: "{publicUrl}/api/v1"
    description: A family's own instance
    variables:
      publicUrl:
        default: https://coop.example
...
security:
  - parentSession: []
paths:
...
  /web/link:
    post:
      operationId: createWebDeviceLink
      tags: [child-session]
      summary: Create a short-lived browser device handoff
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [deviceName]
              properties:
                deviceName: { type: string, maxLength: 80 }
      responses:
        "201":
          description: Created
          headers:
            Cache-Control: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WebDeviceLink" }
        "400":
          description: The device name is invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          description: Too many browser links were created from this address
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /web/link/{linkId}/redeem:
    parameters:
      - $ref: "#/components/parameters/LinkId"
      - $ref: "#/components/parameters/WebLinkSecret"
    post:
      operationId: redeemWebDeviceLink
      tags: [child-session]
      summary: Exchange an approved handoff for an HttpOnly browser session
      security: []
      responses:
        "200":
          description: Linked
          headers:
            Set-Cookie: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ChildProfile" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    parentSession:
      type: http
      scheme: bearer
      description: A parent session token from the login flow.
    childToken:
      type: http
      scheme: bearer
      description: |
        A device token from pairing. Scoped to one child, read-mostly, and
        revocable from the parent app.
    webChildSession:
      type: apiKey
      in: cookie
      name: __Host-coop-child
      description: A Secure, HttpOnly, SameSite=Strict browser device session.
  parameters:
...
    LinkId:
      name: linkId
      in: path
      required: true
      schema: { type: string, format: uuid }
    WebLinkSecret:
      name: X-Coop-Link-Secret
      in: header
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: |
        No such resource, or it is outside the caller's scope. Scope failures
        are reported as 404 so the response does not confirm existence.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    ChildProfile:
      type: object
      required: [id, name, shortsEnabled, watchPageAutoplay, channelDiscoveryEnabled, webLinkingEnabled, allowSelfUnpair]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        avatarId: { type: string }
        shortsEnabled: { type: boolean }
        watchPageAutoplay: { type: boolean }
        videoSearchTiles: { type: boolean }
        channelDiscoveryEnabled: { type: boolean }
        webLinkingEnabled: { type: boolean }
        allowSelfUnpair: { type: boolean }
...
    WebDeviceLink:
      type: object
      required: [id, approvalToken, redemptionToken, expiresAt, qrPayload]
      properties:
        id: { type: string, format: uuid }
        approvalToken: { type: string }
        redemptionToken: { type: string }
        expiresAt: { type: string, format: date-time }
        qrPayload: { type: string }
...